requirements.txt
streamlit==1.45.1
weaviate-client==4.14.4
requests==2.32.3
pandas==2.2.3
